@@ -122,7 +122,7 @@ namespace Udemy.Ecommerce.Infraestructure.Repository
         {
             using (var connection = _connectionFactory.GetConnection)
             {
-                var query = "CustomersGetAll";
+                var query = "CustomersList";
                 var customers = await connection.QueryAsync<Customer>(query, commandType: CommandType.StoredProcedure);
                 return customers;
             }
